Revert "fix: rename common file"
This reverts commit e674408f5142d927cd0f5c057c70b377c22fe9f7.
